+++ src/bin/inx-chronicle/api/routes.rs
@@ -9,14 +9,10 @@ use axum::{
     routing::{get, post},
     Extension, Json, TypedHeader,
 };
-use chronicle::{
-    db::{collections::MilestoneCollection, MongoDb},
-    types::stardust::milestone::MilestoneTimestamp,
-};
+use chronicle::db::MongoDb;
 use hyper::StatusCode;
 use regex::RegexSet;
 use serde::Deserialize;
-use time::{Duration, OffsetDateTime};
 
 use super::{
     auth::Auth,
@@ -29,10 +25,6 @@ use super::{
 };
 
 const ALWAYS_AVAILABLE_ROUTES: &[&str] = &["/health", "/login", "/routes"];
-
-// Similar to Hornet, we enforce that the latest known milestone is newer than 5 minutes. This should give Chronicle
-// sufficient time to catch up with the node that it is connected too. The current milestone interval is 5 seconds.
-const STALE_MILESTONE_DURATION: Duration = Duration::minutes(5);
 
 pub fn routes() -> Router {
     #[allow(unused_mut)]
@@ -88,12 +80,6 @@ pub fn password_verify(
     Ok(hash == argon2::hash_raw(password, salt, &config)?)
 }
 
-fn is_new_enough(timestamp: MilestoneTimestamp) -> bool {
-    // Panic: The milestone_timestamp is guaranteeed to be valid.
-    let timestamp = OffsetDateTime::from_unix_timestamp(timestamp.0 as i64).unwrap();
-    OffsetDateTime::now_utc() <= timestamp + STALE_MILESTONE_DURATION
-}
-
 async fn list_routes(
     ListRoutesQuery { depth }: ListRoutesQuery,
     Extension(config): Extension<ApiData>,
@@ -127,33 +113,13 @@ async fn list_routes(
     Ok(RoutesResponse { routes })
 }
 
-pub async fn is_healthy(database: &MongoDb) -> ApiResult<bool> {
-    #[cfg(feature = "stardust")]
-    {
-        let newest = match database
-            .collection::<MilestoneCollection>()
-            .get_newest_milestone()
-            .await?
-        {
-            Some(last) => last,
-            None => return Ok(false),
-        };
-
-        if !is_new_enough(newest.milestone_timestamp) {
-            return Ok(false);
-        }
-    }
-
-    Ok(true)
-}
-
 pub async fn health(database: Extension<MongoDb>) -> StatusCode {
-    let handle_error = |ApiError { error, .. }| {
+    let handle_error = |error| {
         tracing::error!("An error occured during health check: {error}");
         false
     };
 
-    if is_healthy(&database).await.unwrap_or_else(handle_error) {
+    if database.is_healthy().await.unwrap_or_else(handle_error) {
         StatusCode::OK
     } else {
         StatusCode::SERVICE_UNAVAILABLE
